pass test/loader/get_load_units.test.ts
